separated bse class into its own file
